refactor: move the resource cap onto JsonBudget (#105)
Resources.bounded_json sat public only so VFS could reach it, and its body
was one delegating line, so VFS reached through Resources for a concern
JsonBudget owns. Both callers now use JsonBudget.for_resource. The budget
specs move to their own file with it.
